feat(Default handler): Add showStack option
The `defaultHandler` displays log events on `stderr`. If `stderr` is a TTY device then it displays the log event  formatted for human consumption (if not then as JSON). For `error` events long stack traces can be shown - these can be noisy and distracting, and inappropriate when the error is not an exception e.g. a user entering a bad CLI option.

This commit adds the `showStack` option which defaults to `false`. Note that this does not apply to non-TTY JSON output. It can be turned on easily by apps during debug mode e.g. 

```js
replaceHandlers(data =>
  defaultHandler(data, {
    maxLevel: debug ? LogLevel.debug : LogLevel.info,
    showStack: debug
  })
)
```